+++ trs/core.py
@@ -21,7 +21,7 @@ class PersonYearCombination(object):
         'target',
         'turnover',
         'overbooked',
-        'left_to_book',
+        'left_to_book_external',
         'well_booked',
         'booked_internal',
         'booked_external',
@@ -33,6 +33,7 @@ class PersonYearCombination(object):
         'all_booked_hours',
         'to_book_this_year',
         'all_bookings_percentage',
+        'left_to_turn_over',
     ]
 
     def __init__(self, person, year=None):
@@ -40,7 +41,7 @@ class PersonYearCombination(object):
         if year is None:
             year = datetime.date.today().year
         self.year = year
-        version = 22
+        version = 24
         self.cache_key = 'pycdata-%s-%s-%s-%s' % (
             person.id, person.cache_indicator, year, version)
         has_cached_data = self.get_cache()
@@ -125,19 +126,26 @@ class PersonYearCombination(object):
             well_booked_this_year = booked_this_year.get(id, 0) - overbooked_this_year
             turnover = well_booked_this_year * hourly_tariff[id]
             left_to_book = max(0, (budget[id] - booked_till_now))
+            left_to_turn_over = left_to_book * hourly_tariff[id]
             if is_internal[id]:
                 booked_internal = booked_this_year.get(id, 0)
                 booked_external = 0
+                left_to_book_external = 0
+                # ^^^ TODO: later on we might want to deal with internal
+                # projects that are in fact proper projects and count their
+                # left-to-book hours.
             else:
                 booked_internal = 0
                 booked_external = booked_this_year.get(id, 0)
+                left_to_book_external = left_to_book
 
             project_info = {
                 'booked': booked,
                 'overbooked': overbooked_this_year,
                 'well_booked': well_booked_this_year,
-                'left_to_book': left_to_book,
+                'left_to_book_external': left_to_book_external,
                 'turnover': turnover,
+                'left_to_turn_over': left_to_turn_over,
                 'booked_internal': booked_internal,
                 'booked_external': booked_external,
             }
@@ -181,8 +189,10 @@ class PersonYearCombination(object):
                                 for project in per_project.values()])
         self.turnover = sum([project['turnover']
                              for project in per_project.values()])
-        self.left_to_book = sum([project['left_to_book']
-                                 for project in per_project.values()])
+        self.left_to_turn_over = sum([project['left_to_turn_over']
+                                      for project in per_project.values()])
+        self.left_to_book_external = sum(
+            [project['left_to_book_external'] for project in per_project.values()])
         self.booked_internal = sum([project['booked_internal']
                                     for project in per_project.values()])
         self.booked_external = sum([project['booked_external']
@@ -231,7 +241,7 @@ class PersonYearCombination(object):
 
     @cached_property
     def relative_target_percentage(self):
-        """Return target percentage relative to the elapsed weeks."""
+        """Return target percentage relative to the elapsed days."""
         if self.year == datetime.date.today().year:
             days_elapsed = datetime.date.today().timetuple().tm_yday
             portion_of_year = days_elapsed / 365
@@ -240,10 +250,6 @@ class PersonYearCombination(object):
         if not self.target:  # Division by zero.
             return 100
         return round(self.turnover / self.target * 100 / portion_of_year)
-
-    @cached_property
-    def left_to_turn_over(self):
-        return max((self.target - self.turnover), 0)
 
 
 def get_pyc(person, year=None):